ADD browserify livereload & refactor npm scripts
